Organize storm.rs under bw_1161
