+++ game/src/bw_1161/storm.rs
@@ -4,7 +4,26 @@ use std::mem;
 
 use lazy_static::lazy_static;
 
+use crate::bw;
 use crate::windows;
+
+#[repr(C)]
+pub struct SCode {
+    pub whatever: [u8; 0x4c],
+    pub code_offsets: [*mut u8; 0xa1],
+}
+
+whack_hooks!(stdcall, 0x15000000,
+    0x1503DE90 => InitializeSnpList();
+    0x150380A0 => UnloadSnp(u32);
+);
+
+whack_vars!(init_vars, 0x15000000,
+    0x1505E630 => snp_list_initialized: u32;
+    // Not actually a full entry, just next/prev pointers
+    0x1505AD6C => snp_list: bw::SnpListEntry;
+    0x1505EC04 => surface_copy_code: *mut SCode;
+);
 
 lazy_static! {
     static ref SERR_GET_LAST_ERROR: unsafe extern "stdcall" fn() -> u32 = unsafe {
