+++ game/src/snp.rs
@@ -306,7 +306,7 @@ unsafe extern "stdcall" fn receive_packet(
         *addr = null_mut();
         *data = null_mut();
         *length = 0;
-        crate::storm::SErrSetLastError(STORM_ERROR_NO_MESSAGES_WAITING);
+        bw::with_bw(|bw| bw.storm_set_last_error(STORM_ERROR_NO_MESSAGES_WAITING));
         0
     }
 }
